pyproject.toml
[tool.poetry]
name = "raindrop-labels"
version = "0.1.0"
description = "Automated system for processing raindrop.io bookmarks"
authors = ["dorschre"]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
python-slugify = "^8.0.4"
keybert = "^0.8.5"
openai = "^1.53.0"
markdownify = "^0.13.1"
anthropic = "^0.42.0"
youtube-transcript-api = "^0.6.3"
raindrop-io-py = {git = "https://github.com/ReneDorsch/raindrop-io-py.git"}
fastapi = "^0.115.6"
streamlit = "^1.41.1"
requests = "^2.32.3"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
